requirements.txt
aiohttp
aiofiles
aioshutil
anytree
apscheduler
aria2p
asyncio
beautifulsoup4
bencoding
cloudscraper
dnspython
feedparser
flask
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
gunicorn
git+https://github.com/zevtyardt/lk21.git
httpx
lxml
motor
natsort
pillow
psutil
pybase64
pymongo
pyrogram==2.0.59
python-dotenv
python-magic
qbittorrent-api
requests
telegraph
tenacity
tgcrypto
uvloop
xattr
yt-dlp
